test: split ir tests into modules
